DBReader::getAminoAcidDBSize returns the actual number of sequence residues or profile columns.
Previously it would add 2 bytes per entry for sequences (\n\0) or for 22 additional bytes per column for profiles (and another \0 at the end).

This only has an effect on the E-value computation of profile searches that get swapped (slice search and target profile search), iterative profile search are not affected by the 22 additonal bytes.