breaking: fix parameter singleton in subprojects
With subprojects who have their own Parameter class it can happen that
the MMseqs2 Parameters object is constructed first (in MMseqsBase.cpp)
and then the subproject singleton is accessed and but is already wrongly initialized 
as the parent.

Here we introduce a function initParameterSingleton that decides who to initialize.
Subprojects need to provide this function that calls the correct Parameters constructor now.